Mod/Points moved from float -> double

+++ src/Mod/Points/App/Points.cpp
@@ -70,8 +70,8 @@ Data::Segment* PointKernel::getSubElement(const char* Type, unsigned long n) con
 
 void PointKernel::transformGeometry(const Base::Matrix4D &rclMat)
 {
-    std::vector<Base::Vector3f>& kernel = getBasicPoints();
-    for (std::vector<Base::Vector3f>::iterator it = kernel.begin(); it != kernel.end(); ++it)
+    std::vector<Base::Vector3d>& kernel = getBasicPoints();
+    for (std::vector<Base::Vector3d>::iterator it = kernel.begin(); it != kernel.end(); ++it)
         *it = rclMat * (*it);
 }
 
@@ -94,7 +94,7 @@ void PointKernel::operator = (const PointKernel& Kernel)
 
 unsigned int PointKernel::getMemSize (void) const
 {
-    return _Points.size() * sizeof(Base::Vector3f);
+    return _Points.size() * sizeof(Base::Vector3d);
 }
 
 void PointKernel::Save (Base::Writer &writer) const
@@ -111,8 +111,8 @@ void PointKernel::SaveDocFile (Base::Writer &writer) const
     Base::OutputStream str(writer.Stream());
     uint32_t uCt = (uint32_t)size();
     str << uCt;
-    // store the data without transforming it and save as float, not double
-    for (std::vector<Base::Vector3f>::const_iterator it = _Points.begin(); it != _Points.end(); ++it) {
+    // store the data without transforming it
+    for (std::vector<Base::Vector3d>::const_iterator it = _Points.begin(); it != _Points.end(); ++it) {
         str << it->x << it->y << it->z;
     }
 }
@@ -141,6 +141,9 @@ void PointKernel::RestoreDocFile(Base::Reader &reader)
     str >> uCt;
     _Points.resize(uCt);
     for (unsigned long i=0; i < uCt; i++) {
+// if doubleFileVersion
+//      double x, y, z
+// else
         float x, y, z;
         str >> x >> y >> z;
         _Points[i].Set(x,y,z);
@@ -176,7 +179,7 @@ void PointKernel::getFaces(std::vector<Base::Vector3d> &Points,std::vector<Facet
 // ----------------------------------------------------------------------------
 
 PointKernel::const_point_iterator::const_point_iterator
-(const PointKernel* kernel, std::vector<Base::Vector3f>::const_iterator index)
+(const PointKernel* kernel, std::vector<Base::Vector3d>::const_iterator index)
   : _kernel(kernel), _p_it(index)
 {
     if(_p_it != kernel->_Points.end())
